+++ crazyBird-admin/src/main/java/com/admin/controller/lost/model/LostTypeItem.java
@@ -1,12 +1,13 @@
 package com.admin.controller.lost.model;
 
 public class LostTypeItem {
-	private Integer typeId;
+	private Long typeId;
 	private String typeName;
-	public Integer getTypeId() {
+
+	public Long getTypeId() {
 		return typeId;
 	}
-	public void setTypeId(Integer typeId) {
+	public void setTypeId(Long typeId) {
 		this.typeId = typeId;
 	}
 	public String getTypeName() {
